skills/community/clawhub/r/relive/requirements.txt
conformer==0.3.2
diffusers<0.30.0
gdown
hyperpyyaml>=1.2.0
hydra-core
inflect
jinja2>=3.1.0
librosa>=0.10.0
lightning
matplotlib
modelscope>=1.9.0
numpy==1.26.4
omegaconf>=2.3.0
onnxruntime
openai-whisper
pandas
pillow
pyarrow
pyworld
pyyaml>=6.0
rank_bm25>=0.2.2
requests>=2.28.0
soundfile>=0.12.0
torch==2.3.1
torchaudio==2.3.1
transformers==4.51.3
wget
x_transformers